backend/requirements.txt
flask
flask-cors
openai
python-dotenv
flask_sqlalchemy
gunicorn
